+++ packages/convex-client/src/index.ts
@@ -1,5 +1,6 @@
 export { Convex } from './convex.js';
+export { KeyPair, type Hex, type IKeyPair } from './KeyPair.js';
 export * from './types.js';
-export * from './crypto.js'; 
+export * from './crypto.js';
 export * from './identicon.js';
 export * from './keystore.js';